// src/api/types/repo_status.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BranchInfo } from "./branch_info";
import type { StatusItem } from "./status_item";

export type RepoStatus = { branch: BranchInfo, staged: Array<StatusItem>, unstaged: Array<StatusItem>, conflicts: Array<StatusItem>, };