// src/api/types/status_item.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FileStatusKind } from "./file_status_kind";

export type StatusItem = { path: string, orig_path: string | null, status: FileStatusKind, is_staged: boolean, is_conflicted: boolean, };