// src/api/types/file_diff.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DiffHunk } from "./diff_hunk";
import type { FileStatusKind } from "./file_status_kind";

export type FileDiff = { old_path: string | null, new_path: string, status: FileStatusKind, is_binary: boolean, old_mode: string | null, new_mode: string | null, hunks: Array<DiffHunk>, total_additions: number, total_deletions: number, is_truncated: boolean, total_lines: number, raw_size_bytes: number, };